refactor: update component modules
- update component modules to use new types
- enforce key as required field within the frontend
- ensure global type definitions are only visible or used by the local
  type definition
- TG-216
