+++ frontend/src/components/rubric-builder/CriteriaInput.tsx
@@ -1,4 +1,4 @@
-import React, {
+import {
   ChangeEvent,
   MouseEvent as ReactMouseEvent,
   ReactElement,
